update router, add Oauth, setup new routes

+++ client/src/utilities/fetch.ts
@@ -11,9 +11,8 @@ interface ErrorResponse {
   message: string;
 }
 
-const getToken = async () => {
-  // const token = await AsyncStorage.getItem('duckit-token');
-  const token = 'token';
+const getToken = (): string | null => {
+  const token = sessionStorage.getItem('fullstack_todos') || null;
   return token;
 };
 
@@ -26,38 +25,47 @@ const privateHeaders = (token: string) => ({
   Authorization: `Bearer ${token}`,
 });
 
-export const GET = async <T>(endpoint: string, isProtected = false): Promise<SuccessResponse<T> | ErrorResponse> =>
-  await getToken().then(async (token) => {
-    const response = await fetch(`${BASE_URL}${endpoint}`, {
-      method: 'GET',
-      headers: isProtected ? privateHeaders(token) : publicHeaders,
-    });
-    return await response.json();
-  });
+export const GET = async <T>(endpoint: string, isProtected = false): Promise<SuccessResponse<T> | ErrorResponse> => {
+  const token = getToken();
 
-export const PATCH = async <T>(endpoint: string, body: T): Promise<SuccessResponse | ErrorResponse> =>
-  getToken().then((token) =>
-    fetch(`${BASE_URL}${endpoint}`, {
-      method: 'PATCH',
-      headers: privateHeaders(token),
-      body: JSON.stringify(body),
-    })
-      .then((response) => response.json())
-      .catch((error) => {
-        throw new Error(error.message || 'An error occurred');
-      }),
-  );
+  if (isProtected && !token) throw new Error('No token found');
+
+  const response = await fetch(`${BASE_URL}${endpoint}`, {
+    method: 'GET',
+    headers: isProtected ? privateHeaders(token!) : publicHeaders,
+  });
+  return await response.json();
+};
+
+export const PATCH = async <T>(endpoint: string, body: T): Promise<SuccessResponse | ErrorResponse> => {
+  const token = getToken();
+
+  if (!token) throw new Error('No token found');
+
+  return fetch(`${BASE_URL}${endpoint}`, {
+    method: 'PATCH',
+    headers: privateHeaders(token),
+    body: JSON.stringify(body),
+  })
+    .then((response) => response.json())
+    .catch((error) => {
+      throw new Error(error.message || 'An error occurred');
+    });
+};
 
 export const POST = async <T>(
   endpoint: string,
   body?: T,
   isProtected = false,
 ): Promise<SuccessResponse | ErrorResponse> => {
-  const token = await getToken();
+  const token = getToken();
+
+  if (isProtected && !token) throw new Error('No token found');
+
   try {
     const payload = {
       method: 'POST',
-      headers: isProtected ? privateHeaders(token) : publicHeaders,
+      headers: isProtected ? privateHeaders(token!) : publicHeaders,
     };
 
     if (body) {
@@ -65,6 +73,7 @@ export const POST = async <T>(
     }
 
     const response = await fetch(`${BASE_URL}${endpoint}`, payload);
+
     const data = await response.json();
 
     if (!data.success) {
@@ -77,30 +86,35 @@ export const POST = async <T>(
   }
 };
 
-export const PUT = <T>(endpoint: string, body: T, isProtected = false): Promise<SuccessResponse | ErrorResponse> =>
-  getToken().then((token) =>
-    fetch(`${BASE_URL}${endpoint}`, {
+export const PUT = async <T>(endpoint: string, body: T, isProtected = false): Promise<SuccessResponse | ErrorResponse> => {
+  const token = getToken();
+
+  if (isProtected && !token) throw new Error('No token found');
+
+  try {
+    const response = await fetch(`${BASE_URL}${endpoint}`, {
       method: 'PUT',
-      headers: isProtected ? privateHeaders(token) : publicHeaders,
+      headers: isProtected ? privateHeaders(token!) : publicHeaders,
       body: JSON.stringify(body),
-    })
-      .then((response) => response.json())
-      .then((data) => {
-        if (!data.success) {
-          throw new Error(data.message || 'Request failed');
-        }
+    });
+    const data = await response.json();
+    if (!data.success) {
+      throw new Error(data.message || 'Request failed');
+    }
+    return data;
+  } catch (error: any) {
+    throw new Error(error.message || 'An error occurred');
+  }
+};
 
-        return data;
-      })
-      .catch((error) => {
-        throw new Error(error.message || 'An error occurred');
-      }),
-  );
+export const DELETE = async (endpoint: string): Promise<SuccessResponse | ErrorResponse> => {
+  const token = getToken();
 
-export const DELETE = (endpoint: string): Promise<SuccessResponse | ErrorResponse> =>
-  getToken().then((token) =>
-    fetch(`${BASE_URL}${endpoint}`, {
-      method: 'DELETE',
-      headers: privateHeaders(token),
-    }).then((response) => response.json()),
-  );
+  if (!token) throw new Error('No token found');
+
+  const response = await fetch(`${BASE_URL}${endpoint}`, {
+    method: 'DELETE',
+    headers: privateHeaders(token),
+  });
+  return await response.json();
+};
